refactor: 스타일 컴포넌트 파일 styled로 지정된 파일들 styles로 변경
