+++ src/components/common-ui/star-rating/StarRating.styles.ts
