+++ src/components/movies/Movies.styles.ts
