+++ src/components/series/Series.styles.ts
