+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/globalstyle.ts","./src/components/index.ts","./src/components/commonui/posterbox/posterbox.styled.ts","./src/components/common-ui/error/error.styles.ts","./src/components/common-ui/header/header.styles.ts","./src/components/common-ui/loader/loader.styles.ts","./src/components/edit-profile/editprofileform.styles.ts","./src/components/movies/movies.styled.ts","./src/components/not-found/notfound.styles.ts","./src/components/series/series.styled.ts","./src/components/sign-in/signinform.styles.ts","./src/components/sign-up/signupform.styles.ts","./src/constants/comment.ts","./src/constants/user.ts","./src/hooks/useerrorhandler.ts","./src/hooks/mutations/useeditprofile.ts","./src/hooks/mutations/usesignin.ts","./src/hooks/mutations/usesignout.ts","./src/hooks/mutations/usesignup.ts","./src/hooks/mutations/usedeactivateaccount.ts","./src/hooks/queries/usecheckduplicate.ts","./src/hooks/queries/usefetchuser.ts","./src/pages/index.ts","./src/schemas/user/baseuserschema.ts","./src/schemas/user/editprofileschema.ts","./src/schemas/user/signinschema.ts","./src/schemas/user/signupschema.ts","./src/stores/userstore.ts","./src/types/auth.ts","./src/types/components.ts","./src/types/error.ts","./src/types/user.ts","./src/utils/isapierror.ts","./src/app.tsx","./src/main.tsx","./src/components/commonui/posterbox/posterbox.tsx","./src/components/commonui/skeleton/skeleton.tsx","./src/components/common-ui/error/error.tsx","./src/components/common-ui/header/header.tsx","./src/components/common-ui/loader/loader.tsx","./src/components/edit-profile/editprofileform.tsx","./src/components/not-found/notfound.tsx","./src/components/sign-in/signinform.tsx","./src/components/sign-up/signupform.tsx","./src/layout/defaultlayout.tsx","./src/pages/editprofile.tsx","./src/pages/home.tsx","./src/pages/movies.tsx","./src/pages/notfoundpage.tsx","./src/pages/series.tsx","./src/pages/signin.tsx","./src/pages/signup.tsx","./src/routes/router.tsx","./vite-env.d.ts"],"version":"5.7.2"}
+{"root":["./src/globalstyle.ts","./src/components/index.ts","./src/components/otheruserprofile/otheruserprofile.styles.ts","./src/components/common-ui/button/button.styles.ts","./src/components/common-ui/input/input.styles.ts","./src/components/common-ui/profile/profile.styles.ts","./src/components/common-ui/error/error.styles.ts","./src/components/common-ui/header/header.styles.ts","./src/components/common-ui/loader/loader.styles.ts","./src/components/common-ui/poster/posterbox.styles.ts","./src/components/common-ui/star-rating/starrating.styles.ts","./src/components/edit-profile/editprofileform.styles.ts","./src/components/home-main/homemain.styles.ts","./src/components/movies/movies.styles.ts","./src/components/not-found/notfound.styles.ts","./src/components/series/series.styles.ts","./src/components/sign-in/signinform.styles.ts","./src/components/sign-up/signupform.styles.ts","./src/constants/comment.ts","./src/constants/user.ts","./src/hooks/useerrorhandler.ts","./src/hooks/mutations/useeditprofile.ts","./src/hooks/mutations/usesignin.ts","./src/hooks/mutations/usesignout.ts","./src/hooks/mutations/usesignup.ts","./src/hooks/mutations/usedeactivateaccount.ts","./src/hooks/queries/usecheckduplicate.ts","./src/hooks/queries/usefetchuser.ts","./src/pages/index.ts","./src/schemas/user/baseuserschema.ts","./src/schemas/user/editprofileschema.ts","./src/schemas/user/signinschema.ts","./src/schemas/user/signupschema.ts","./src/stores/userstore.ts","./src/types/auth.ts","./src/types/button.ts","./src/types/components.ts","./src/types/error.ts","./src/types/input.ts","./src/types/profile.ts","./src/types/user.ts","./src/utils/isapierror.ts","./src/app.tsx","./src/main.tsx","./src/components/otheruserprofile/otheruserprofile.tsx","./src/components/common-ui/button/button.tsx","./src/components/common-ui/input/input.tsx","./src/components/common-ui/profile/profile.tsx","./src/components/common-ui/error/error.tsx","./src/components/common-ui/header/header.tsx","./src/components/common-ui/loader/loader.tsx","./src/components/common-ui/poster/posterbox.tsx","./src/components/common-ui/skeleton/skeleton.tsx","./src/components/common-ui/star-rating/starrating.tsx","./src/components/edit-profile/editprofileform.tsx","./src/components/home-main/homemain.tsx","./src/components/not-found/notfound.tsx","./src/components/sign-in/signinform.tsx","./src/components/sign-up/signupform.tsx","./src/layout/defaultlayout.tsx","./src/pages/editprofile.tsx","./src/pages/home.tsx","./src/pages/movies.tsx","./src/pages/notfoundpage.tsx","./src/pages/otheruserprofilepage.tsx","./src/pages/series.tsx","./src/pages/signin.tsx","./src/pages/signup.tsx","./src/routes/router.tsx","./vite-env.d.ts"],"errors":true,"version":"5.7.2"}
